Move all errors to separate file
