+++ src/metadata.rs
@@ -3,5 +3,5 @@ mod track_id;
 mod values;
 
 pub use self::metadata::Metadata;
-pub use self::track_id::{InvalidTrackID, TrackID};
+pub use self::track_id::TrackID;
 pub use self::values::MetadataValue;
